success rate now integer

+++ app/src/main/java/com/michaelcarrano/seven_min_workout/data/TimeExercise.java
@@ -3,7 +3,7 @@ package com.michaelcarrano.seven_min_workout.data;
 public class TimeExercise extends ExerciseStats {
     private boolean completedLastTime = false;
     private int totalCompleted = 0;
-    private double completedPercentage = 0;
+    private int completedPercentage = 0;
     private boolean currentStatus = false;
 
     public TimeExercise(String exerciseName) {
@@ -18,11 +18,11 @@ public class TimeExercise extends ExerciseStats {
         this.completedLastTime = completedLastTime;
     }
 
-    public double getCompletedPercentage() {
+    public int getCompletedPercentage() {
         return completedPercentage;
     }
 
-    public void setCompletedPercentage(double completedPercentage) {
+    public void setCompletedPercentage(int completedPercentage) {
         this.completedPercentage = completedPercentage;
     }
